+++ src/bin/designTokens.ts
@@ -2,11 +2,11 @@
 
 import parserRuntime from 'yargs-parser';
 import designTokens from '@/index.ts';
-import { logError, configFileParser } from '@/functions';
+import { configFileParser, logError } from '@/functions';
 import fs from 'fs';
 
 const args = parserRuntime(process.argv.slice(2));
 
 configFileParser(args, fs)
-  .then(config => designTokens(args, config))
+  .then(config => designTokens(config))
   .catch(err => logError(err));
