+++ src/index.ts
@@ -1,10 +1,9 @@
-import { buildStyleDictionary } from '@/styleDictionary/StyleDictionary.ts';
-import { Arguments } from 'yargs-parser';
+import { buildStyleDictionary } from '@/styleDictionary/styleDictionary.ts';
 import { ParseConfig } from '@/types/designTokens';
 import { figmaApiConnection } from '@/api';
 import { createJsonTokenFiles, EMOJIS, log } from '@/functions';
 
-const designTokens = (args: Arguments, config: ParseConfig) => {
+const designTokens = (config: ParseConfig) => {
   const { settings } = config;
 
   figmaApiConnection(settings).then(async generatedTokens => {
